+++ CHANGELOG.md
@@ -1,5 +1,6 @@
 ## [16.8.2]
 - [ContentControl][iOS] If content is CollectionView we wrap it in a grid with row=auto.
+- [ListItem] Options now inherit the BindingContext of the ListItem.
 
 ## [16.8.1]
 - [Chip] Made sure background color of chip is set correctly.